Server-FastAPI_Supabase/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
supabase>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
httpx>=0.24.0
requests>=2.31.0
email-validator>=2.1.0
mangum>=0.17.0